Treemap
Reduced the maximum fontsize of the treemap labels from 40px to 25px,
as per design. Apart from aligning the fontsize with the design i also
modified the x position of the treemap labels.

Note: the modification of the x position of the treemap label only
relates to the top text element, not the bottom one which contains
the percentage.

Signed-off-by: Jim Fung <jim@zimmermanzimmerman.nl>